# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Real-time voting API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: polls
    description: Poll management
  - name: votes
    description: Vote operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /polls:
    get:
      operationId: listPolls
      tags: [polls]
      summary: List all polls
      responses:
        "200":
          description: List of polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"
    post:
      operationId: createPoll
      tags: [polls]
      summary: Create a new poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollInput"
      responses:
        "201":
          description: Created poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /polls/stats:
    get:
      operationId: getPollStats
      tags: [polls]
      summary: Get aggregate voting statistics
      responses:
        "200":
          description: Poll statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollStats"

  /polls/{pollId}:
    get:
      operationId: getPoll
      tags: [polls]
      summary: Get a single poll with live vote counts
      parameters:
        - name: pollId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Poll with vote counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePoll
      tags: [polls]
      summary: Delete a poll
      parameters:
        - name: pollId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /polls/{pollId}/votes:
    post:
      operationId: castVote
      tags: [votes]
      summary: Cast a vote (accepted into in-memory queue, 202 response)
      parameters:
        - name: pollId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "202":
          description: Vote accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteAccepted"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Double-voting detected — this voter has already cast a ballot on this poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many vote attempts — rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    PollOption:
      type: object
      required: [id, text, votes]
      properties:
        id:
          type: string
        text:
          type: string
        votes:
          type: integer

    Poll:
      type: object
      required: [id, question, options, totalVotes, createdAt]
      properties:
        id:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        totalVotes:
          type: integer
        createdAt:
          type: string
          format: date-time

    PollInput:
      type: object
      required: [question, options]
      properties:
        question:
          type: string
          minLength: 3
        options:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 10

    VoteInput:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: string

    VoteAccepted:
      type: object
      required: [message, queued]
      properties:
        message:
          type: string
        queued:
          type: boolean

    PollStats:
      type: object
      required: [totalPolls, totalVotes]
      properties:
        totalPolls:
          type: integer
        totalVotes:
          type: integer
        mostVotedPoll:
          type: ["object", "null"]
          properties:
            id:
              type: string
            question:
              type: string
            totalVotes:
              type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string